update: restructure the install packages
